requirements.txt
Flask==2.2.5
tensorflow==2.10.0
opencv-python==4.7.0.72
numpy==1.21.6
scikit-learn==1.0.2
joblib==1.1.1
albumentations==1.2.1
Pillow==9.3.0
Werkzeug==2.2.3
pyswarms==1.3.0
matplotlib==3.5.3
pandas==1.3.5 